+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 
 - Implement edX open response assessment events pydantic models
 - Implement edx peer instruction events pydantic models
+- Implement xAPI VideoDownloaded pydantic model
+(using xAPI TinCan `downloaded` verb)
 
 ### Changed
 
@@ -32,7 +34,7 @@ and this project adheres to
 - Clean xAPI pydantic models naming convention
 - Upgrade `fastapi` to `0.97.0`
 - Upgrade `sentry_sdk` to `1.25.1`
-- Set Clickhouse `client_options` to a dedicated Pydantic model
+- Set Clickhouse `client_options` to a dedicated pydantic model
 - Upgrade `httpx` to `0.24.1`
 - Force a valid (JSON-formatted) IFI to be passed for the `/statements`
 GET query `agent` filtering